latest version is using u64 weight (#106)

+++ packages/types/src/interfaces/runtime/types.ts
@@ -351,7 +351,7 @@ export interface TradingPair extends Struct {
 export interface ValidatorId extends AccountId {}
 
 /** @name Weight */
-export interface Weight extends u32 {}
+export interface Weight extends u64 {}
 
 /** @name WeightMultiplier */
 export interface WeightMultiplier extends Fixed64 {}
